concertbitebackend/requirements.txt
asgiref==3.2.10
astroid
bcrypt
beautifulsoup4==4.9.2
certifi==2020.6.20
cffi
dj-database-url==0.5.0
Django==3.1.2
djangorestframework==3.12.1
gunicorn==20.0.4
isort
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.6
pycparser
pylint
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml
whitenoise==5.2.0
wrapt==1.11.2
django-cors-headers==3.5.0